feat: add decimal support for stock and quantity in transactions
- Updated NumpadModal to allow decimal input for quantities.
- Modified Create and Edit product pages to handle decimal values for stock and conversion fields.
- Introduced QtyNominalModal for quantity and nominal input with decimal support.
- Updated database migrations to change stock and quantity fields to decimal type.
- Added tests to ensure decimal quantities can be added and updated in the cart accurately.

+++ app/Models/StockMutation.php
@@ -25,9 +25,9 @@ class StockMutation extends Model
         'id' => 'integer',
         'product_id' => 'integer',
         'reference_id' => 'integer',
-        'qty' => 'integer',
-        'stock_before' => 'integer',
-        'stock_after' => 'integer',
+        'qty' => 'float',
+        'stock_before' => 'float',
+        'stock_after' => 'float',
         'created_by' => 'integer',
     ];
